refactor(desktop): move renderer nav helpers
