+++ apps/desktop/src/main/__tests__/renderer-shell-source-helpers.ts
@@ -11,6 +11,7 @@ const sourcePaths = [
   'app-shell.tsx',
   'cached-theme-bootstrap.ts',
   'conversation-markdown.ts',
+  'nav-selection.ts',
 ] as const;
 
 export const RENDERER_SHELL_SOURCE_REPO_PATHS: readonly string[] = sourcePaths.map(
