+++ apps/desktop/src/renderer/app-shell.tsx
@@ -107,6 +107,7 @@ import {
 import { safeLocalStorageGet, safeLocalStorageSet } from './browser-storage';
 import { applyLocalSessionRead, createSessionListRefresher, type SessionListRefresher, type SessionReadBoundaries } from './session-read-state';
 import { renderConversationMarkdown } from './conversation-markdown';
+import { countSessions, filterSessions, readNavSelection } from './nav-selection';
 
 const NO_REAL_CONNECTION_CODE = 'NO_REAL_CONNECTION';
 const NO_REAL_CONNECTION_REASON_RE = /NO_REAL_CONNECTION:([a-z_]+): /;
@@ -3553,33 +3554,6 @@ const modeDescriptions: Record<PermissionMode, string> = {
   bypass: '跳过全部工具确认，包括破坏性操作、特权操作和浏览器操作。',
 };
 
-function readNavSelection(): NavSelection {
-  try {
-    const raw = safeLocalStorageGet('maka-nav-selection-v1');
-    if (!raw) return { section: 'sessions', filter: 'chats' };
-    const parsed = JSON.parse(raw) as { section?: string; filter?: string };
-    // PR-SIDEBAR-IA-0 Phase 2 fixup (xuan `94c7bf0f`): fail-closed.
-    // `'search'` was briefly a `NavSelection.section` during the
-    // Phase 2 initial commit; the fixup removes it because `搜索`
-    // is now a modal trigger, not a section. An older localStorage
-    // entry with `{section:'search'}` would otherwise leave the
-    // app stuck on an invalid section. Reject anything that is not
-    // in the current closed-enum.
-    if (parsed.section === 'skills') return { section: 'skills' };
-    if (parsed.section === 'automations') return { section: 'automations' };
-    if (parsed.section === 'daily-review') return { section: 'daily-review' };
-    if (
-      parsed.section === 'sessions' &&
-      (parsed.filter === 'chats' || parsed.filter === 'flagged' || parsed.filter === 'archived')
-    ) {
-      return parsed as NavSelection;
-    }
-  } catch {
-    /* fall through */
-  }
-  return { section: 'sessions', filter: 'chats' };
-}
-
 function readSessionListWidth(): number {
   const stored = Number(safeLocalStorageGet('maka-chat-list-width-v1'));
   if (Number.isFinite(stored) && stored > 0) return clampSessionListWidth(stored);
@@ -3673,24 +3647,4 @@ function clamp(value: number, min: number, max: number): number {
 
 function clampSessionListWidth(value: number): number {
   return Math.round(clamp(value, SESSION_LIST_EXPANDED_MIN_WIDTH, SESSION_LIST_EXPANDED_MAX_WIDTH));
-}
-
-function filterSessions(sessions: SessionSummary[], selection: NavSelection): SessionSummary[] {
-  const filter = selection.section === 'sessions' ? selection.filter : 'chats';
-  switch (filter) {
-    case 'flagged':
-      return sessions.filter((session) => session.isFlagged && !session.isArchived && session.lastMessageAt);
-    case 'archived':
-      return sessions.filter((session) => session.isArchived);
-    case 'chats':
-      return sessions.filter((session) => !session.isArchived && session.lastMessageAt);
-  }
-}
-
-function countSessions(sessions: SessionSummary[]) {
-  return {
-    chats: sessions.filter((session) => !session.isArchived && session.lastMessageAt).length,
-    flagged: sessions.filter((session) => session.isFlagged && !session.isArchived && session.lastMessageAt).length,
-    archived: sessions.filter((session) => session.isArchived).length,
-  };
 }
